new file:   include/partial_vision.h 	modified:   include/prey.h 	modified:   src/blind.cpp 	modified:   src/hexaworld.cpp 	new file:   src/partial_vision.cpp 	modified:   src/prey.cpp

+++ src/blind.cpp
@@ -20,7 +20,7 @@ void Blind::save_expected_reward(Prey_state_action &sa, double reward, uint32_t 
     double aging = ((double)_buffer[index]->visits) / ((double)_buffer[index]->visits + 1);
     double step_reward = reward * ( pow(_config.discount,episode_steps - sa.iteration) ) + _config.step_cost * sa.iteration;
     _buffer[index]->reward = aging * _buffer[index]->reward + step_reward * ( 1 - aging);
-    (double)_buffer[index]->visits ++;
+    _buffer[index]->visits ++;
 }
 
 Blind::Blind(World &w, Visibility &vi, Prey_config &config) :
